Prodel example, display positioning fixes, gave up on vertical

+++ core/src/main/java/com/zarbosoft/merman/editor/display/Blank.java
@@ -3,5 +3,5 @@ package com.zarbosoft.merman.editor.display;
 import com.zarbosoft.merman.editor.Context;
 
 public interface Blank extends CourseDisplayNode {
-	void setConverseSpan(Context context, int converse);
+	void setConverseSpan(Context context, double converse);
 }
