+++ core/src/main/java/com/zarbosoft/merman/editor/wall/Bedding.java
@@ -1,10 +1,10 @@
 package com.zarbosoft.merman.editor.wall;
 
 public class Bedding {
-	public final int before;
-	public final int after;
+	public final double before;
+	public final double after;
 
-	public Bedding(final int before, final int after) {
+	public Bedding(final double before, final double after) {
 		this.before = before;
 		this.after = after;
 	}
