+++ core/src/test/java/com/zarbosoft/merman/editor/display/MockeryDisplayNode.java
@@ -1,15 +1,15 @@
 package com.zarbosoft.merman.editor.display;
 
 public abstract class MockeryDisplayNode implements DisplayNode {
-  public int converse;
+  public double converse;
 
   @Override
-  public int converse() {
+  public double converse() {
     return converse;
   }
 
   @Override
-  public void setConverse(final int converse, final boolean animate) {
+  public void setConverse(final double converse, final boolean animate) {
     this.converse = converse;
   }
 }
